HW_47/requirements.txt
pyodbc
python-dotenv
